# sessao_2/crud_go/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GoAPI",
    "description": "CRUD de processos",
    "version": "0.1.0"
  },
  "paths": {
    "/processos": {
      "post": {
        "summary": "Criar",
        "description": "Create: cadastra um processo e devolve 201 com o header Location.",
        "operationId": "criar_processos_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessoEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Processo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar",
        "description": "Read: lista os processos; ?uf=SP filtra pelo estado.",
        "operationId": "listar_processos_get",
        "parameters": [
          {
            "name": "uf",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uf"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Processo"
                  },
                  "title": "Response Listar Processos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/processos/{processo_id}": {
      "get": {
        "summary": "Buscar",
        "description": "Read: um processo pelo id.",
        "operationId": "buscar_processos__processo_id__get",
        "parameters": [
          {
            "name": "processo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Processo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Processo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Substituir",
        "description": "Update completo: troca todos os campos.",
        "operationId": "substituir_processos__processo_id__put",
        "parameters": [
          {
            "name": "processo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Processo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessoEntrada"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Processo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Atualizar",
        "description": "Update parcial: troca só os campos enviados no body.",
        "operationId": "atualizar_processos__processo_id__patch",
        "parameters": [
          {
            "name": "processo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Processo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessoParcial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Processo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remover",
        "description": "Delete: remove o processo e responde 204, sem body.",
        "operationId": "remover_processos__processo_id__delete",
        "parameters": [
          {
            "name": "processo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Processo Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Processo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "cnj": {
            "type": "string",
            "title": "Cnj"
          },
          "uf": {
            "type": "string",
            "title": "Uf"
          },
          "comarca": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comarca"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cnj",
          "uf"
        ],
        "title": "Processo",
        "description": "Um processo cadastrado."
      },
      "ProcessoEntrada": {
        "properties": {
          "cnj": {
            "type": "string",
            "title": "Cnj"
          },
          "uf": {
            "type": "string",
            "title": "Uf"
          },
          "comarca": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comarca"
          }
        },
        "type": "object",
        "required": [
          "cnj",
          "uf"
        ],
        "title": "ProcessoEntrada",
        "description": "Corpo de POST e PUT: todos os campos."
      },
      "ProcessoParcial": {
        "properties": {
          "cnj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnj"
          },
          "uf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uf"
          },
          "comarca": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comarca"
          }
        },
        "type": "object",
        "title": "ProcessoParcial",
        "description": "Corpo de PATCH: só os campos que mudam."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}